Refactored header buttons to multiple files
